fix: Move OnboardingPropTypes and MobileRouter to be transpiled
This reverts commit 8c774117c4ea8dba3c95cf0bc6308254c4fa7680.
