+++ packages/cozy-authentication/package.json
@@ -28,6 +28,7 @@
     "parcel": "^1.10.3",
     "react": "^16.6.3",
     "react-dom": "^16.6.3",
+    "react-router": "3.2.1",
     "react-styleguidist": "6.5.3",
     "stylus": "^0.54.5"
   },
