+++ packages/cozy-authentication/src/MobileRouter.jsx
@@ -1,8 +1,8 @@
 import React, { Component } from 'react'
 import { Router, withRouter } from 'react-router'
 import Proptypes from 'prop-types'
-import Authentication from './src/Authentication'
-import Revoked from './src/Revoked'
+import Authentication from './Authentication'
+import Revoked from './Revoked'
 import {
   readState,
   secretExchange,
@@ -12,7 +12,7 @@ import {
   clearState,
   checkIfOnboardingLogin,
   addProtocolToURL
-} from './src/utils/onboarding'
+} from './utils/onboarding'
 import {
   onboardingInformationsPropTypes,
   onboardingPropTypes
