+++ packages/cozy-authentication/src/MobileRouter.spec.jsx
