+++ packages/cozy-authentication/src/OnboardingPropTypes.js
