+++ packages/cozy-authentication/src/steps/SelectServer.jsx
@@ -12,7 +12,7 @@ import 'cozy-ui/assets/icons/ui/next.svg'
 import 'cozy-ui/assets/icons/ui/lock.svg'
 import styles from '../styles.styl'
 import { ButtonLinkRegistration } from './ButtonLinkRegistration'
-import { onboardingPropTypes } from '../../OnboardingPropTypes'
+import { onboardingPropTypes } from '../OnboardingPropTypes'
 
 require('url-polyfill')
